Flask/requirements.txt
click==8.1.8
colorama==0.4.6
Flask==2.2.5
importlib-metadata==6.7.0
itsdangerous==2.1.2
jinja2==3.1.6
MarkupSafe==2.1.5
typing-extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
PyJWT==2.8.0
requests==2.31.0
Flask-Cors==5.0.0